// src/views/LHD/components/HistoricalChart/types.ts
export enum DatasetNames {
  // $ Values
  MarketCap = 'Market Cap',
  TotalExtractableLiquidity = 'Total Extractable Liq',
  OwnedLiquidity = 'Owned Extractable Liq',
  LiquidityDebt = 'Liquidity Debt',
  // Scores
  TotalScore = 'Total Score',
  HealthScore = 'Strength Score',
  OwnershipScore = 'Ownership Score',
  ConcentrationScore = 'Concentration Score',
}

export enum HistoricalDataType {
  Dollar = 'Dollars',
  Score = 'Scores',
  Liquidity = 'Liquidity',
}